requirements.txt
# Python 3.10.x
gradio==5.29.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
matplotlib==3.10.3
numpy==1.26.2
openai==1.79.0
pandas==2.2.3
python-dotenv==1.1.0
seaborn==0.13.2
SQLAlchemy==2.0.41
typer==0.15.4
pydantic==2.11.4